+++ sources/packages/backend/apps/db-migrations/src/sql/DynamicFormConfigurations/Rollback-add-authorization-key-column.sql
@@ -1,0 +1,2 @@
+ALTER TABLE
+    sims.dynamic_form_configurations DROP COLUMN authorization_key;
